Docker container and CMake fixes
This adds a Docker container. As reported by Bharat, the docker
container would need boost python, but this dependency isn't
needed by aartfaac2ms, so the cmake conf was changed to remove
the boost python dependency. Also, Tammo Jan noted that cfitsio
is not properly included, which is also fixed.